Enum plane_split::Intersection

pub enum Intersection<T> {
    Coplanar,
    Outside,
    Inside(T),
}

Polygon intersection results.

Variants
Coplanar

Polygons are coplanar, including the case of being on the same plane.

Outside

Polygon planes are intersecting, but polygons are not.

Inside(T)

Polygons are actually intersecting.

Methods

impl<T> Intersection<T>

pub fn is_outside(&self) -> bool

Return true if the intersection is completely outside.

pub fn is_inside(&self) -> bool

Return true if the intersection cuts the source polygon.