Struct plane_split::BspSplitter

pub struct BspSplitter<A: Copy> {
    result: Vec<Polygon<A>>,
    nodes: Vec<BspNode>,
    polygons: Vec<Polygon<A>>,
}

Binary Space Partitioning splitter, uses a BSP tree.

Fields
result: Vec<Polygon<A>>nodes: Vec<BspNode>polygons: Vec<Polygon<A>>
Implementations

impl<A: Copy> BspSplitter<A>

pub fn new() -> Self

Create a new BSP splitter.

impl<A> BspSplitter<A>
where A: Copy + Debug + Default,

pub fn reset(&mut self)

Put the splitter back in it initial state.

Call this at the beginning of every frame when reusing the splitter.

pub fn add(&mut self, poly: Polygon<A>)

Add a polygon to the plane splitter.

This is where most of the expensive computation happens.

pub fn sort(&mut self, view: Vector3D<f64>) -> &[Polygon<A>]

Sort the added and split polygons against the view vector.

Call this towards the end of the frame after having added all polygons.

pub fn solve( &mut self, input: &[Polygon<A>], view: Vector3D<f64>, ) -> &[Polygon<A>]
where A: Copy,

Process a set of polygons at once.

fn insert(&mut self, node_idx: NodeIdx, value: &Polygon<A>)

Insert a value into the sub-tree starting with this node. This operation may spawn additional leafs/branches of the tree.

pub fn order(&self, node: NodeIdx, base: &Polygon<A>, out: &mut Vec<Polygon<A>>)

Build the draw order of this sub-tree into an out vector, so that the contained planes are sorted back to front according to the view vector defined as the base plane front direction.