Struct plane_split::Polygon

pub struct Polygon<T, U, A> {
    pub points: [Point3D<T, U>; 4],
    pub plane: Plane<T, U>,
    pub anchor: A,
}

A convex polygon with 4 points lying on a plane.

Fields
points: [Point3D<T, U>; 4]

Points making the polygon.

plane: Plane<T, U>

A plane describing polygon orientation.

anchor: A

A simple anchoring index to allow association of the produced split polygons with the original one.

Implementations

impl<T, U, A> Polygon<T, U, A> where
T: Copy + Debug + ApproxEq<T> + Sub<T, Output = T> + Add<T, Output = T> + Mul<T, Output = T> + Div<T, Output = T> + Zero + One + Float,
U: Debug,
A: Copy,

pub fn from_points(points: [Point3D<T, U>; 4], anchor: A) -> Option<Self>

Construct a polygon from points that are already transformed. Return None if the polygon doesn't contain any space.

pub fn from_rect(rect: Rect<T, U>, anchor: A) -> Self

Construct a polygon from a non-transformed rectangle.

pub fn from_transformed_rect<V>(
rect: Rect<T, V>,
transform: Transform3D<T, V, U>,
anchor: A
) -> Option<Self> where
T: Trig + Neg<Output = T>,

Construct a polygon from a rectangle with 3D transform.

pub fn from_transformed_rect_with_inverse<V>(
rect: Rect<T, V>,
transform: &Transform3D<T, V, U>,
inv_transform: &Transform3D<T, U, V>,
anchor: A
) -> Option<Self> where
T: Trig + Neg<Output = T>,

Construct a polygon from a rectangle with an invertible 3D transform.

pub fn untransform_point(&self, point: Point3D<T, U>) -> Point2D<T>

Bring a point into the local coordinate space, returning the 2D normalized coordinates.

pub fn transform<V>(
&self,
transform: &Transform3D<T, U, V>
) -> Option<Polygon<T, V, A>> where
T: Trig,
V: Debug,

Transform a polygon by an affine transform (preserving straight lines).

pub fn is_valid(&self) -> bool

Check if all the points are indeed placed on the plane defined by the normal and offset, and the winding order is consistent.

pub fn is_empty(&self) -> bool

Check if the polygon doesn't contain any space. This may happen after a sequence of splits, and such polygons should be discarded.

pub fn contains(&self, other: &Self) -> bool

Check if this polygon contains another one.

pub fn project_on(&self, vector: &Vector3D<T, U>) -> LineProjection<T>

Project this polygon onto a 3D vector, returning a line projection. Note: we can think of it as a projection to a ray placed at the origin.

pub fn intersect_plane(&self, other: &Plane<T, U>) -> Intersection<Line<T, U>>

Compute the line of intersection with an infinite plane.

pub fn intersect(&self, other: &Self) -> Intersection<Line<T, U>>

Compute the line of intersection with another polygon.

fn split_impl(
&mut self,
first: (usize, Point3D<T, U>),
second: (usize, Point3D<T, U>)
) -> (Option<Self>, Option<Self>)

pub fn split(&mut self, line: &Line<T, U>) -> (Option<Self>, Option<Self>)

👎 Deprecated:
Use split_with_normal instead

Split the polygon along the specified Line. Will do nothing if the line doesn't belong to the polygon plane.

pub fn split_with_normal(
&mut self,
line: &Line<T, U>,
normal: &Vector3D<T, U>
) -> (Option<Self>, Option<Self>)

Split the polygon along the specified Line, with a normal to the split line provided. This is useful when called by the plane splitter, since the other plane's normal forms the side direction here, and figuring out the actual line of split isn't needed. Will do nothing if the line doesn't belong to the polygon plane.